refactor(ai-backend): extract schemas and remove inquiry alias fallback
